+++ slider/src/brain/netApp.cpp
